pyproject.toml
[project]
name = "my-mcp-chromadb"
version = "1.0.0"
description = "My personal MCP server for AI agent long-term memory with ChromaDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0"
]

# Configure UV to prefer CPU-only packages for containers
index-strategy = "unsafe-best-match"
extra-index-url = [
    "https://download.pytorch.org/whl/cpu"  # PyTorch CPU-only index
]

[tool.uv.sources]
# Force CPU-only PyTorch packages when available
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }